requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
django-extensions==3.1.5
ipython==8.4.0
mysqlclient==2.1.0
PyJWT==2.4.0
requests==2.27.1
